Make the harness model the 32-bit millis() wrap faithfully
Asked what happens when the device has been up 169 days -- 3.4 millis()
wraps -- and the SSID disappears. Measuring it exposed a flaw in the test
harness, not in the firmware.

The stub declared `unsigned long millis()`, matching the Arduino
signature. On the ESP32-C3 `unsigned long` is 32 bits, so `millis() -
start` wraps and the wrap-safe idioms work. On an x86-64 host it is 64
bits, so the subtraction never wraps and those idioms behave differently
from the target. Every wrap-related conclusion drawn from this harness
was therefore untrustworthy -- and it showed: measuring the relay pulse
across a wrap reported 60 seconds instead of 90, which looked like a
firmware bug and was purely an artifact. The stub now returns uint32_t,
matching the target exactly.

With that fixed the firmware is correct at every point tested: the 90
second pulse stays 90 seconds when millis() wraps during it, during the
ten-minute RESET_DELAY, and mid-way through the failure sequence; the
whole recovery still takes 16.5 minutes and ends in the usual one-hour
sleep. That follows from the code -- every timing uses unsigned
subtraction -- but it had never actually been exercised.

New tests MW1-MW3 pin it down: a wrap during the relay pulse, during
RESET_DELAY and mid-sequence, and a run at 169 days (1716698112 ms, three
wraps behind). Verified they are meaningful by rebuilding against a copy
of the old 64-bit stub, where MW1 fails.

504/504 checks pass in 156 scenarios, on the IDF 5 and IDF 6 branches and
under ASan+UBSan, no compiler warnings, coverage 98.13%.

+++ test/stubs/main_stub.cpp
@@ -127,7 +127,7 @@ void Print::flushLine() {
   buf_.clear();
 }
 
-unsigned long millis() { return g_millis; }
+uint32_t millis() { return g_millis; }   // lasd a megjegyzest az Arduino.h-ban
 void pinMode(uint8_t p, uint8_t m) { (void)m; g_pinState[p] = -1; }
 void digitalWrite(uint8_t p, uint8_t v) {
   if (g_pinState[p] != v) simLog("pin" + std::to_string(p) + "=" + (v ? "HIGH" : "LOW"));
